requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gensim==4.3.2
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pymssql==2.3.0
python-dotenv==1.0.1
scikit-learn==1.2.2
scipy==1.10.0
tzdata==2024.1
Werkzeug==3.0.3
itsdangerous==2.2.0
